Update benchmarks for 0.4.0 (#10)
Apart from adding new configurations for 0.4.0 final, there were a few changes in scripts:

* Migrate to Python 3
* Runner script records cold compilation time
* Binaries are stored on the side and can be reused across script invocations
* New summary scripts to summarize compilation time and binary size across configurations

+++ scripts/run.py
@@ -4,6 +4,7 @@ import os
 import errno
 import subprocess as subp
 import shutil as sh
+import time
 
 def mkdir(path):
     try:
@@ -35,10 +36,13 @@ def run(cmd):
     return subp.check_output(cmd)
 
 def compile(bench, compilecmd):
+    start = time.time_ns()
     cmd = [sbt, '-J-Xmx6G', 'clean']
     cmd.append('set mainClass in Compile := Some("{}")'.format(bench))
     cmd.append(compilecmd)
-    return run(cmd)
+    run(cmd)
+    end = time.time_ns()
+    return end - start
 
 sbt = where('sbt')
 
@@ -62,10 +66,10 @@ benchmarks = [
 ]
 
 configurations = [
-        # 'jvm',
-        # 'scala-native-0.3.8',
-        # 'baseline',
-        'current'
+        'scala-native-0.3.9',
+        'scala-native-0.4.0-M2',
+        'scala-native-0.4.0-2.11',
+        'scala-native-0.4.0-2.12',
 ]
 
 if 'GRAALVM_HOME' in os.environ:
@@ -75,7 +79,7 @@ if 'GRAALVM_HOME' in os.environ:
     ]
 
 runs = 20
-batches = 4000
+batches = 2000
 batch_size = 1
 
 if __name__ == "__main__":
@@ -83,32 +87,49 @@ if __name__ == "__main__":
         for bench in benchmarks:
             print('--- conf: {}, bench: {}'.format(conf, bench))
 
+            resultsdir = os.path.join('results', conf, bench)
+            mkdir(resultsdir)
+
+            bindir = os.path.join("binaries", conf, bench)
+            mkdir(bindir)
+
+            binfile = os.path.join(bindir, bench)
+
             input = slurp(os.path.join('input', bench))
             output = slurp(os.path.join('output', bench))
             compilecmd = slurp(os.path.join('confs', conf, 'compile'))
             runcmd = slurp(os.path.join('confs', conf, 'run')).replace('$BENCH', bench).replace('$HOME', os.environ['HOME']).split(' ')
 
-            if os.path.exists(os.path.join('confs', conf, 'build.sbt')):
-                sh.copyfile(os.path.join('confs', conf, 'build.sbt'), 'build.sbt')
-            else:
-                os.remove('build.sbt')
+            if not os.path.exists(binfile):
+                if os.path.exists(os.path.join('confs', conf, 'build.sbt')):
+                    sh.copyfile(os.path.join('confs', conf, 'build.sbt'), 'build.sbt')
+                else:
+                    os.remove('build.sbt')
 
-            if os.path.exists(os.path.join('confs', conf, 'plugins.sbt')):
-                sh.copyfile(os.path.join('confs', conf, 'plugins.sbt'), 'project/plugins.sbt')
-            else:
-                os.remove('project/plugins.sbt')
+                if os.path.exists(os.path.join('confs', conf, 'build.properties')):
+                    sh.copyfile(os.path.join('confs', conf, 'build.properties'), 'project/build.properties')
+                else:
+                    os.remove('project/build.properties')
 
-            compile(bench, compilecmd)
-            resultsdir = os.path.join('results', conf, bench)
-            mkdir(resultsdir)
+                if os.path.exists(os.path.join('confs', conf, 'plugins.sbt')):
+                    sh.copyfile(os.path.join('confs', conf, 'plugins.sbt'), 'project/plugins.sbt')
+                else:
+                    os.remove('project/plugins.sbt')
 
-            for n in xrange(runs):
+                compile_time = compile(bench, compilecmd)
+
+                with open(os.path.join(resultsdir, "compile_time"), 'w') as f:
+                    f.write(str(compile_time))
+
+                sh.move(runcmd[0], binfile)
+
+            for n in range(runs):
                 print('--- run {}/{}'.format(n, runs))
 
-                cmd = []
-                cmd.extend(runcmd)
+                cmd = [binfile]
+                cmd.extend(runcmd[1:])
                 cmd.extend([str(batches), str(batch_size), input, output])
                 out = run(cmd)
-                with open(os.path.join(resultsdir, str(n)), 'w+') as resultfile:
+                with open(os.path.join(resultsdir, str(n)), 'wb') as resultfile:
                     resultfile.write(out)
